erltity behaviour (erltity v0.0.1)

Types

cursor/0

-type cursor() :: binary() | undefined.

db_driver/0

-type db_driver() :: epgsql.

find_opts/0

-type find_opts() ::
          #{page => non_neg_integer(), page_size => non_neg_integer(), offset => non_neg_integer()}.

opts/0

-type opts() :: #{host => string(), username => string(), password => string(), database => string()}.

path/0

-type path() :: binary().

register_opts/0

-type register_opts() :: #{save_module => path(), save_create_table => path()}.

schemas_path/0

-type schemas_path() :: binary().

Callbacks

create/1

-callback create(CreateRequest) -> Result
                    when
                        CreateRequest :: map(),
                        Result :: {ok, Entity} | {error, Reason},
                        Entity :: map(),
                        Reason :: term().

delete/1

-callback delete(Id) -> Result when Id :: term(), Result :: ok | {error, Reason}, Reason :: term().

find/1

-callback find(Id) -> Result
                  when
                      Id :: term(),
                      Result :: {ok, Entity} | {error, Reason},
                      Entity :: map(),
                      Reason :: term().

find/2

-callback find(Filters, Opts) -> Result
                  when
                      Filters :: map(),
                      Opts :: find_opts(),
                      Result :: {ok, [Entity]} | {error, Reason},
                      Entity :: map(),
                      Reason :: term().

update/2

-callback update(Id, UpdateRequest) -> Result
                    when
                        UpdateRequest :: map(),
                        Id :: term(),
                        Result :: {ok, Entity} | {error, Reason},
                        Entity :: map(),
                        Reason :: term().

Functions

create(EntityName, CreateRequest)

-spec create(EntityName, CreateRequest) -> Result
                when
                    EntityName :: atom(),
                    CreateRequest :: map(),
                    Result :: {ok, Entity} | {error, Reason},
                    Entity :: map(),
                    Reason :: term().

Saves a new entity of type EntityName with the given CreateRequest.

delete(EntityName, Id)

-spec delete(EntityName, Id) -> Result
                when
                    EntityName :: atom(), Id :: term(), Result :: ok | {error, Reason}, Reason :: term().

Deletes an existing entity of type EntityName with the given Id.

find(EntityName, Id)

-spec find(EntityName, Id) -> Result
              when
                  EntityName :: atom(),
                  Id :: term(),
                  Result :: {ok, Entity} | {error, Reason},
                  Entity :: map(),
                  Reason :: term().

Finds an entity of type EntityName with the given Id.

find(EntityName, Filters, Opts)

-spec find(EntityName, Filters, Opts) -> Result
              when
                  EntityName :: atom(),
                  Filters :: map(),
                  Opts :: find_opts(),
                  Result :: {ok, [Entity]} | {error, Reason},
                  Entity :: map(),
                  Reason :: term().

Finds entities of type EntityName that match the given Filters and Opts, Opts can include pagination options like page, page_size, and offset.

register(SchemaPath, Opts)

-spec register(SchemaPath, Opts) -> Result
                  when
                      SchemaPath :: binary(),
                      Opts :: register_opts(),
                      Result :: ok | {error, Reason},
                      Reason :: term().

Registers a schema from the given path and generates the new Erlang module.

start_link(DbDriver, Opts)

-spec start_link(DbDriver, Opts) -> Result
                    when DbDriver :: db_driver(), Opts :: opts(), Result :: supervisor:startlink_ret().

Starts the erltity application connecting to the database using the given driver and options.

stop(StopArg)

-spec stop(StopArg) -> Result when StopArg :: term(), Result :: ok.

Stops the erltity application.

update(EntityName, Id, UpdateRequest)

-spec update(EntityName, Id, UpdateRequest) -> Result
                when
                    EntityName :: atom(),
                    UpdateRequest :: map(),
                    Id :: term(),
                    Result :: {ok, Entity} | {error, Reason},
                    Entity :: map(),
                    Reason :: term().

Updates an existing entity of type EntityName with the given Id and UpdateRequest.