erltity (erltity v0.0.1)

Types

db_driver/0

-type db_driver() :: epgsql.

opts/0

-type opts() :: #{host => string(), username => string(), password => string(), database => string()}.

path/0

-type path() :: binary().

register_opts/0

-type register_opts() :: #{save_module => path(), save_create_table => path()}.

schemas_path/0

-type schemas_path() :: binary.

Functions

register(SchemaPath, Opts)

-spec register(SchemaPath, Opts) -> Result
                  when
                      SchemaPath :: binary(),
                      Opts :: register_opts(),
                      Result :: ok | {error, Reason},
                      Reason :: term().

Registers a schema from the given path and generates the new Erlang module.

start_link(DbDriver, Opts)

-spec start_link(DbDriver, Opts) -> Result
                    when DbDriver :: db_driver(), Opts :: opts(), Result :: supervisor:startlink_ret().

Starts the erltity application connecting to the database using the given driver and options.

stop(StopArg)

-spec stop(StopArg) -> Result when StopArg :: term(), Result :: ok.

Stops the erltity application.